chore: update naming from login to home
